+++ ChangeLog.md
@@ -8,6 +8,8 @@
   `FileTrustStore` implementation to retrieve all stored host and port
   combinations from the trust store.
   ([#22](https://github.com/davep/wasat/pull/22))
+- Added `trust_store` public property to the `Client` class to expose the underlying trust store.
+  ([#22](https://github.com/davep/wasat/pull/22))
 
 ## v0.4.0